feat: migrate zsh config away from omz (#8)
Major refactor of ZSH configuration to begin migrating away from OMZ. Interoperability is achieved via OMZ's plugin concept and associated plugin overrides; albiet the non-omz plugin loader is significantly more simplistic. Only thing not included as of yet is the inclusion of loading plugin completions but that's because I don't have any as of yet.